requirements.txt
Flask>=2.0
twilio>=7.0
python-dotenv>=0.19
requests>=2.25